feat: function calls with args
This introduces a buggy implementation of calling functions with
arguments. The arguments will be passed over register if possible, but
after r4-r7 are filled they will be placed on the stack.

The current implementation takes advantage of the stack too much for
many operations and emits a lot of instructions. There is no function
inlining, and so the cost of a function is increased due to the IC10
environment's limit of 128 lines. Additionally the arguments passed via
function do not seem to load their values correctly. Further work must
improve upon this before it is useful.