fix(compiler): Differentiate between register and stack location for locals
Previously all locals were assumed to have a position on the stack. When
calling a function they are passed via register most commonly and this
assumption was causing the values to fail to load.

I'm not 100% convinced that this is the best method to do this, but for
now it works.