chore(examples): rewrite greenhouse regulator
The previous example was based around one for MScript. A fundamental
limitation in MScript is that write instructions cannot accept constant
values. This caused the program to use a variable in order to write.

The program will not compile under the current version of ayysee,
but the intent is for the next changes to enable the functionality that
this program needs in order to function.