feat: Constant value definitions
Constant values are ones that do not allow for mutation. They are
intended to remove the need for magic numbers, especially when specified
in multiple locations in code.

This implementation allows for constants to be redefined later in the
program. This isn't a specific design choice but rather the decision to
implement a check will be made at a later date.