feat(cli): add compilation output type
Usage\: add the `--output` or `-o` flag to a compiler invocation to
allow users to select what type of output. The default remains to
compile to Stationeers MIPS, but now there is the ability to specify if
you only want an AST. In the future this could be used to display any
intermediate representation.